fix type deduction

+++ extras/rapidfuzz_amalgamated.hpp
@@ -1,7 +1,7 @@
 //  Licensed under the MIT License <http://opensource.org/licenses/MIT>.
 //  SPDX-License-Identifier: MIT
 //  RapidFuzz v1.0.2
-//  Generated: 2023-10-09 03:12:47.555069
+//  Generated: 2023-10-09 03:30:56.866043
 //  ----------------------------------------------------------
 //  This file is an amalgamation of multiple different files.
 //  You probably shouldn't edit it directly.
@@ -5611,14 +5611,14 @@ void jaro_similarity_simd(Range<double*> scores, const detail::BlockPatternMatch
         auto s2_cur = s2;
 
         int64_t lastRelevantChar = 0;
-        int64_t maxBound = 0;
+        ptrdiff_t maxBound = 0;
         unroll<int, vec_width>([&](auto i) {
             int64_t s1_len = s1_lengths[result_index + i];
             int64_t Bound = jaro_bounds(s1_len, s2_cur.size());
 
             if (s1_len + Bound > lastRelevantChar) lastRelevantChar = s1_len + Bound;
 
-            if (Bound > maxBound) maxBound = Bound;
+            if (Bound > maxBound) maxBound = static_cast<ptrdiff_t>(Bound);
 
             boundMaskSize_[i] = bit_mask_lsb<VecType>(static_cast<int>(2 * Bound));
             boundMask_[i] = bit_mask_lsb<VecType>(static_cast<int>(Bound + 1));
